Use "id" instead of "testsuite"
Using "testsuite" as a command-line parameter was confusing, because a single
output.xml can contain multiple Robot Framework test suites. So, what "id"
(previously "testsuite") is referring to now is basically an instance of
output.xml. The "id" parameter is mainly intended for internal use in Prometheus
as it does not map to anything in Robot Framework itself.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>